+++ lib/mediawiki/query.rb
@@ -129,6 +129,64 @@ module MediaWiki
             return false
           end
         end
+
+        # Gets a hash-of-arrays containing all the groups the user can add and
+        #   remove people from.
+        # @return [Boolean/Hash] False if get_current_user_meta is false, else
+        #   a hash containing arrays of all the groups the user can add/remove
+        #   people from.
+        def get_changeable_groups
+          response = get_current_user_meta('changeablegroups')
+          if response != false
+            ret = {}
+            add = []
+            remove = []
+            addself = []
+            removeself = []
+            changeablegroups = response['query']['userinfo']['changeablegroups']
+            puts changeablegroups
+            changeablegroups['add'].each { |g| puts g; add.push(g) }
+            changeablegroups['remove'].each { |g| puts g; remove.push(g) }
+            changeablegroups['add-self'].each { |g| puts g; addself.push(g) }
+            changeablegroups['remove-self'].each { |g| puts g; removeself.push(g) }
+            ret['add'] = add
+            ret['remove'] = remove
+            ret['addself'] = addself
+            ret['removeself'] = removeself
+            return ret
+          else
+            return false
+          end
+        end
+
+        # Gets the currently logged in user's real name.
+        # @return [String/Nil] Nil if they don't have a real name set, or their
+        #   real name.
+        def get_realname
+          response = get_current_user_meta('realname')
+          if response['query']['userinfo']['realname'] == ''
+            return nil
+          else
+            return response['query']['userinfo']['realname']
+          end
+        end
+
+        # Gets the currently logged in user's email address.
+        # @return [String/Nil] Nil if their email is not set, or their email.
+        def get_email_address
+          response = get_current_user_meta('email')
+          if response['query']['userinfo']['email'] == ''
+            return nil
+          else
+            return response['query']['userinfo']['email']
+          end
+        end
+
+        def get_current_user_options
+          response = get_current_user_meta('options')
+          ret = {}
+          response['query']['userinfo']['options'].each { |k, v| ret[k] = v }
+        end
       end
     end